+++ lib/dartapi_core.dart
@@ -7,12 +7,15 @@ export 'src/core/base_controller.dart';
 export 'src/core/logger.dart';
 export 'src/core/serializable.dart';
 export 'src/core/global_exception_handler.dart';
+export 'src/core/background_task.dart';
+export 'src/core/websocket_route.dart';
 
 export 'src/openapi/security_scheme.dart';
 export 'src/openapi/openapi_generator.dart';
 export 'src/openapi/docs_controller.dart';
 
 export 'src/utils/utils.dart';
+export 'src/utils/multipart.dart';
 
 export 'src/middleware/rate_limit_middleware.dart';
 export 'src/middleware/request_id_middleware.dart';